requirements.txt
Fabric==1.9.0
Flask==0.10.1
Flask-Admin==1.0.8
Flask-MongoRest==0.1.1
Flask-Views==0.2.1
Flask-WTF==0.9.5
Jinja2==2.7.1
MarkupSafe==0.18
Pillow==2.4.0
WTForms==1.0.5
Werkzeug==0.9.4
blinker==1.3
ecdsa==0.11
flask-mongoengine==0.7.0
itsdangerous==0.23
mimerender==0.5.4
mongoengine==0.8.7
paramiko==1.14.0
pycrypto==2.6.1
pymongo==2.7.1
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.3
wsgiref==0.1.2
